util written in ts
